third_party/blink/web_tests/accessibility/selection-change-notification-textarea.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>

<textarea autofocus id="textarea">textarea</textarea>

<script>
'use strict';

async_test((t) => {
  // This forces building the accessibility tree, because selection change
  // events only fire on elements that already exist.
  accessibilityController.accessibleElementById('dummy');

  const element = document.getElementById('textarea');
  const axElement = accessibilityController.accessibleElementById('textarea');

  axElement.setNotificationListener(t.step_func((notification, intents) => {
    if (notification == 'SelectedTextChanged') {
      assert_array_equals(intents,
                          ['AXEventIntent(setSelection,character,forward)']);
      axElement.unsetNotificationListener();
      t.done();
    }
  }));

  element.setSelectionRange(0, 1);
});

</script>
